saveConversation in memory class
